pyproject.toml
[project]
name = "foundation-models-benchmark"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9",
  "pyyaml>=6.0",
  "numpy",
  "torch>=2.0",
  "torchvision",
  "scipy",
  "pandas",
  "matplotlib",
  "seaborn",
  "scikit-learn",
  "datasets",
  "huggingface_hub",
  "transformers",
  "timm",
  "einops",
  "astropy",
  "FrEIA",
  "umap-learn",
  "tqdm",
]

[project.scripts]
fmb = "fmb.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
